Clarify the react guide.
When I first read line 41, I thought that it was saying that only the directly-following code example should go into the `app.jsx` file, so I was confused as to where I should put the rest of the code.  I know that there aren't any other dumb people out there like me, so this is just for the next time I read through the guide.  :)