Get feathers-hooks-common doc to appear in GitBook TOC
File hooks/common_hooks.md does not appear in the TOC
even though its in SUMMARY.md. No one has provided an explaination.
The file name is the only one containing an underscore.
Renaming the file to a simple name to see if that changes
the TOC. SUMMARY.md changed.