Demonstrate how to change the config file location (#1098)
I had to spend a bit of time hunting this down from a dependency’s documentation. Adding it here so others won’t have to.

If there is a more direct way to achieve the same result with the @feathersjs/configuration API directly, please document that instead. Looking at the code, I couldn’t find it.