Fixed typo in "basics/real-time" #what's next
From:
In this chapter we added Socket.io transport and saw how to use channels to send events from server to client. In the next chapter we will look how to use browser Feathers application and client services to handle this events easily.

To:
In this chapter we added the Socket.io transport and saw how to use channels to send events from server to client. In the next chapter we will look at how to use the browser Feathers application and client services to handle these events easily.