Change storageKey value to feathers-jwt in examle
I was just trying to run the code with ra-data-feathers but it didn't work. After debugging, I discovered that the @feathersjs/authentication-client setup script that I copy/paste from this doc is using a different field name.

`auth` is used in the docs of @feathersjs/authentication-client, in the library itself default is `feathers-jwt`, and  `token` in the ra-data-feathers. it's a complete mess :(

Since new users often run the code by copying examples from the documentation, in order not to create unnecessary obstacles and problems for other new users, I think it will be useful to add this fix (I also made a request to the ra-data-feathers to correct the value for the same).