Replace Jade with Pug as example template engine
Jade project was renamed to Pug in 2015.  This page seems really old to use Jade as an example template engine.

I've added link to Pug website, mentioned path to pug template file that would be used and link to Express guide on using template engines that also happens to use Pug as example view template engine.